mint removed from extensions and moved to storage
