+++ packages/tokens/fungible/src/extensions/capped/storage.rs
@@ -20,8 +20,6 @@ pub const CAP_KEY: Symbol = symbol_short!("CAP");
 ///
 /// * We recommend using this function in the constructor of your smart
 ///   contract.
-/// * Cap functionality is designed to be used in conjunction with the
-///   `mintable` extension.
 pub fn set_cap(e: &Env, cap: i128) {
     if cap < 0 {
         panic_with_error!(e, FungibleTokenError::InvalidCap);
