+++ packages/tokens/fungible/src/extensions/mod.rs
@@ -1,4 +1,3 @@
 pub mod burnable;
 pub mod capped;
 pub mod metadata;
-pub mod mintable;
