+++ packages/tokens/fungible/src/storage.rs
@@ -1,7 +1,7 @@
 use soroban_sdk::{contracttype, panic_with_error, Address, Env};
 use stellar_constants::{BALANCE_EXTEND_AMOUNT, BALANCE_TTL_THRESHOLD};
 
-use crate::fungible::{emit_approve, emit_transfer, FungibleTokenError};
+use crate::fungible::{emit_approve, emit_mint, emit_transfer, FungibleTokenError};
 
 /// Storage key that maps to [`AllowanceData`]
 #[contracttype]
@@ -349,4 +349,42 @@ pub fn update(e: &Env, from: Option<&Address>, to: Option<&Address>, amount: i12
         let total_supply = total_supply(e) - amount;
         e.storage().instance().set(&StorageKey::TotalSupply, &total_supply);
     }
+}
+
+/// Creates `amount` of tokens and assigns them to `to`. Updates
+/// the total supply accordingly.
+///
+/// # Arguments
+///
+/// * `e` - Access to the Soroban environment.
+/// * `to` - The address receiving the new tokens.
+/// * `amount` - The amount of tokens to mint.
+///
+/// # Errors
+///
+/// refer to [`update`] errors.
+///
+/// # Events
+///
+/// * topics - `["mint", to: Address]`
+/// * data - `[amount: i128]`
+///
+/// # Security Warning
+///
+/// ⚠️ SECURITY RISK: This function has NO AUTHORIZATION CONTROLS ⚠️
+///
+/// It is the responsibility of the implementer to establish appropriate access
+/// controls to ensure that only authorized accounts can execute minting
+/// operations. Failure to implement proper authorization could lead to
+/// security vulnerabilities and unauthorized token creation.
+///
+/// You probably want to do something like this (pseudo-code):
+///
+/// ```ignore
+/// let admin = read_administrator(e);
+/// admin.require_auth();
+/// ```
+pub fn mint(e: &Env, to: &Address, amount: i128) {
+    update(e, None, Some(to), amount);
+    emit_mint(e, to, amount);
 }
