+++ packages/tokens/fungible/test_snapshots/test/mint_base_implementation_has_no_auth.1.json
