+++ packages/tokens/non-fungible/src/extensions/burnable/test.rs
@@ -5,7 +5,7 @@ extern crate std;
 use soroban_sdk::{contract, testutils::Address as _, Address, Env};
 use stellar_event_assertion::EventAssertion;
 
-use crate::{extensions::mintable::sequential_mint, Base};
+use crate::Base;
 
 #[contract]
 struct MockContract;
@@ -18,7 +18,7 @@ fn burn_works() {
     let owner = Address::generate(&e);
 
     e.as_contract(&address, || {
-        let token_id = sequential_mint(&e, &owner);
+        let token_id = Base::sequential_mint(&e, &owner);
 
         Base::burn(&e, &owner, token_id);
 
@@ -40,7 +40,7 @@ fn burn_from_with_approve_works() {
     let spender = Address::generate(&e);
 
     e.as_contract(&address, || {
-        let token_id = sequential_mint(&e, &owner);
+        let token_id = Base::sequential_mint(&e, &owner);
 
         Base::approve(&e, &owner, &spender, token_id, 1000);
         Base::burn_from(&e, &spender, &owner, token_id);
@@ -64,7 +64,7 @@ fn burn_from_with_operator_works() {
     let operator = Address::generate(&e);
 
     e.as_contract(&address, || {
-        let token_id = sequential_mint(&e, &owner);
+        let token_id = Base::sequential_mint(&e, &owner);
 
         Base::approve_for_all(&e, &owner, &operator, 1000);
 
@@ -88,7 +88,7 @@ fn burn_from_with_owner_works() {
     let owner = Address::generate(&e);
 
     e.as_contract(&address, || {
-        let token_id = sequential_mint(&e, &owner);
+        let token_id = Base::sequential_mint(&e, &owner);
 
         Base::burn_from(&e, &owner, &owner, token_id);
 
@@ -111,7 +111,7 @@ fn burn_with_not_owner_panics() {
     let spender = Address::generate(&e);
 
     e.as_contract(&address, || {
-        let token_id = sequential_mint(&e, &owner);
+        let token_id = Base::sequential_mint(&e, &owner);
 
         Base::burn(&e, &spender, token_id);
     });
@@ -127,7 +127,7 @@ fn burn_from_with_insufficient_approval_panics() {
     let spender = Address::generate(&e);
 
     e.as_contract(&address, || {
-        let token_id = sequential_mint(&e, &owner);
+        let token_id = Base::sequential_mint(&e, &owner);
 
         Base::burn_from(&e, &spender, &owner, token_id);
     });
@@ -143,7 +143,7 @@ fn burn_with_non_existent_token_panics() {
     let non_existent_token_id = 2;
 
     e.as_contract(&address, || {
-        let _token_id = sequential_mint(&e, &owner);
+        let _token_id = Base::sequential_mint(&e, &owner);
 
         Base::burn(&e, &owner, non_existent_token_id);
     });
