+++ packages/tokens/non-fungible/src/extensions/mod.rs
@@ -1,4 +1,3 @@
 pub mod burnable;
 pub mod consecutive;
 pub mod enumerable;
-pub mod mintable;
